Device/requirement.txt
asgiref==3.6.0
Django==4.1.7
django-jazzmin==2.6.0
django-model-utils==1.4.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
pymongo==3.12.3
pytz==2022.7.1
sqlparse==0.2.4
tzdata==2022.7
